requirements.txt
bs4
# csv
# json
requests
# time
faiss-cpu
sentence-transformers
groq
pandas
python-dotenv
streamlit